+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deferred credential response when credential data is not ready is now 202 ACCEPTED
 - Deferred credential transaction_id will not change anymore during deferred flow
 - Added `deferred_credential_endpoint` and `batch_credential_issuance` with min batch size of 10 to sample.compose.yml
+- Prohibit renewal of SUSPENDED and REVOKED VCs and throw Renewal Exception.
 
 ## 2.3.1